working on rebuild exec
