+++ Makefile
@@ -7,12 +7,13 @@ FLAGS := -Wall -Werror -Wextra #-fsanitize=address -g3
 
 SRCS := srcs/parsing/nigga-tests.c \
 		srcs/parsing/parse_redirecs.c \
-		srcs/exec/pipex.c \
-		srcs/exec/fork_process.c \
+		srcs/exec/exec_hub.c \
 		srcs/exec/ft_split.c \
-		srcs/exec/pipe_process.c \
+		srcs/exec/env_checker.c \
 		srcs/exec/pipex_utils.c \
-		srcs/exec/utils.c \
+		srcs/exec/exec_process.c \
+		srcs/exec/sorting_mod.c \
+		srcs/exec/sorting_hub.c \
 
 OBJS := $(SRCS:.c=.o)
