+++ srcs/exec/env_checker.c
@@ -1,0 +1,41 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   env_checker.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: rgeral <rgeral@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/09 18:50:11 by rgeral            #+#    #+#             */
+/*   Updated: 2022/05/09 18:50:24 by rgeral           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../incs/minishell.h"
+
+char	**path(char	**env)
+{
+	int		i;
+	int		j;
+	char	**result;
+	char	*temp;
+
+	i = 0;
+	j = 0;
+	while (env[i])
+	{
+		if (ft_memcmp(env[i], "PATH=", 5) == 0)
+		{
+			result = ft_split(&env[i][5], ':');
+			while (result[j])
+			{
+				temp = result[j];
+				result[j] = ft_strjoin(result[j], "/");
+				free(temp);
+				j++;
+			}
+			return (result);
+		}
+		i++;
+	}
+	return (NULL);
+}
