+++ srcs/exec/pipex_utils.c
@@ -6,7 +6,7 @@
 /*   By: rgeral <rgeral@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/14 11:24:21 by rgeral            #+#    #+#             */
-/*   Updated: 2022/05/09 18:44:33 by rgeral           ###   ########.fr       */
+/*   Updated: 2022/05/09 19:13:03 by rgeral           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -92,44 +92,6 @@ size_t	mod_ft_strlen(const char	*str)
 	return (i);
 }
 
-void	execute(t_args *p, char **args, int nb)
-{
-	char	*tmp;
-	int		j;
-
-	j = 0;
-	int check;
-
-	check = 0;
-	/*
-		Vérifie tout les path + fonction et check si c'est executable, si oui, execute, "sinon invalid path"
-		ex : bin/ls
-	*/
-
-	while (p->path[j])
-	{
-		tmp = ft_strjoin(p->path[j], args[0]);
-		//dprintf(1, "valeur de path[%d] : %s\n", j, p->path[j]);
-		if (access(tmp, F_OK | X_OK) == 0)
-		{
-			check = 1;
-			break ;
-		}
-		j++;
-	}
-	if (access(tmp, F_OK | X_OK) != 0 && nb < p->argc - 1)
-	{
-		perror("Invalid Path");
-		exit(EXIT_FAILURE);
-	}
-	else if (tmp)
-	{
-		dprintf(2, "valeur de tmp : %s\n", tmp);
-		args[0] = tmp;
-		execve(args[0], args, p->env);
-
-	}
-}
 /*
 Making my own dup2 to get error message 
 */
