+++ srcs/exec/sorting_hub.c
@@ -1,0 +1,35 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   sorting_hub.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: rgeral <rgeral@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/09 18:57:33 by rgeral            #+#    #+#             */
+/*   Updated: 2022/05/09 19:41:11 by rgeral           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../incs/minishell.h"
+
+void	one_arg(t_args *d, t_argmode *argv)
+{
+	char	**args;
+	int		argc;
+
+	args = ft_split_len(argv[d->acutal_arg].arg, ' ', &argc);
+	execute(d, args, d->acutal_arg);
+}
+
+void	sorting_hub(t_args *d, t_argmode *argv)
+{
+	if (d->argc < 2)
+		one_arg(d, argv);
+	while (d->acutal_arg < d->argc)
+	{
+		sorting_mod(d, argv);
+		d->j++;
+		d->acutal_arg++;
+	}
+	
+}
